c++ sans warning

+++ tutorials/30-days-of-code/30-more-exceptions.cpp
@@ -20,7 +20,7 @@ public:
         if (n < 0 || p < 0)
             throw invalid_argument("n and p should be non-negative");
 
-        return pow(n, p);
+        return (int) pow(n, p);
     }
 };
